+++ scripts/aggregate-translations/generate-translation-file.js
@@ -17,19 +17,29 @@ const sortMessages = (messages) => {
 };
 
 const translationFile = (locale, messages) => (
-  `import { addLocaleData } from 'react-intl';
-import localeData from 'react-intl/locale-data/${locale.split('-')[0]}';
+  `'use strict';\n
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.messages = exports.locale = exports.areTranslationsLoaded = undefined;
 
-addLocaleData(localeData);
+var _reactIntl = require('react-intl');
 
-const messages = ${JSON.stringify(messages, null, 2)};
-const areTranslationsLoaded = true;
-const locale = '${locale}';
-export {
-  areTranslationsLoaded,
-  locale,
-  messages
-};`);
+var _${locale} = require('react-intl/locale-data/${locale}');
+
+var _${locale}2 = _interopRequireDefault(_${locale});
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+(0, _reactIntl.addLocaleData)(_${locale}2.default);
+
+var messages = ${JSON.stringify(messages, null, 2)};
+var areTranslationsLoaded = true;
+var locale = '${locale}';
+export.areTranslationsLoaded = areTranslationsLoaded;
+export.locale = locale;
+export.messages = messages;
+`);
 
 const generateTranslationFile = (locale, messages) => {
   const sortedMessages = sortMessages(messages);