+++ scripts/aggregate-translations/write-i18n-loaders.js
@@ -3,17 +3,19 @@ const startCase = require('lodash.startcase');
 const supportedLocales = require('./i18nSupportedLocales');
 
 const createIntlLoader = (loaderName, locale) => (
-  `const ${loaderName} = () =>
-require.ensure([],
-  require => require('intl/locale-data/jsonp/${locale}.js'),
-  '${locale}-intl-local');\n
+  `'use strict';\n
+var ${loaderName} = function ${loaderName}() {
+require.ensure([], function (require) {
+  return require('intl/locale-data/jsonp/${locale}.js');
+}, '${locale}-intl-local');\n
 `);
 
 const createTranslationLoader = (loaderName, locale) => (
-  `const ${loaderName} = (callback, scope) => {
-  require.ensure([], (require) => {
+  `'use strict';\n
+var ${loaderName} = function ${loaderName}(callback, scope) {
+  require.ensure([], function (require) {
     // eslint-disable-next-line
-    const i18n = require('./${locale}.js');
+    var i18n = require('./${locale}.js');
     callback.call(scope, i18n);
   }, '${locale}-translations');
 };\n
@@ -43,7 +45,7 @@ const writeLoaders = (type, locales, fileSystem, outputDir) => {
 
   // Create the loader exports
   const loaderString = JSON.stringify(loaders, null, 2).replace(/"/g, '');
-  loaderFile += `const ${type}Loaders = ${loaderString};\n\nmodule.exports = ${type}Loaders;`;
+  loaderFile += `var ${type}Loaders = ${loaderString};\n\nmodule.exports = ${type}Loaders;`;
 
   // Write the loaders file
   const loaderPath = path.resolve(outputDir, `${type}Loaders.js`);